+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 ## 4.64.0 (2025-11-14)
 
 - changed: Upgrade @polkadot/api to latest to support Asset Hub migration
+- removed: Removed 'api.bscscan.com' server due to deprecation
 
 ## 4.63.2 (2025-11-13)